refactor: Tab 컴포넌트 내부 분리 및 Arrow svg 폴더 생성
