+++ zoo-client/src/app/page.tsx
@@ -1,11 +1,9 @@
 import './globals.css'
-import Tab from '@/components/Tab/Tab'
 
 export default function Home() {
   return (
     <div className="figure-m-14 grid min-h-screen grid-rows-[20px_1fr_20px] items-center justify-items-center gap-16 bg-black p-8 pb-20 font-[family-name:var(--font-geist-sans)] sm:p-20">
       Hello Zoo Client
-      <Tab />
     </div>
   )
 }
